chore: move necessary constants and add story
