+++ packages/picasso-query-builder/src/utils/constants.ts
@@ -1,0 +1,4 @@
+export const emptyQueryBuilderQuery = {
+  rules: [],
+  combinator: 'and',
+}
